Add GdsToggleStyle and replace gdsToggleTint
Introduce a new GdsToggleStyle (ToggleStyle) that preserves the native system toggle and applies the GDS tint (.l3Positive01) via .toggleStyle(.gds). Remove the old convenience modifier and deprecated toggle style types (Toggle+GdsModifier.swift, SebGreenToggleStyle.swift) and update usages across the codebase and snapshot tests (ComponentShowcase, RowItem, GreenRow, GdsToggleSnapshotTests) to use .toggleStyle(.gds). This centralizes toggle appearance and removes obsolete compatibility shims.
